first pass moving types for video on a separate folder
