+++ packages/js/src/index.ts
@@ -6,19 +6,17 @@ export type {
   ClientEvents,
   Emitter,
   LayoutEvent,
-  LayoutEvents,
-  MemberEvents,
+  MemberEvent,
   MemberJoinedEventName,
   MemberLeftEventName,
   MemberTalkingEventNames,
   MemberUpdatedEventName,
   RoomEventNames,
   RoomEvent,
-  RoomEvents,
   RoomMember,
   RoomMemberCommon,
   RoomMemberProperties,
-  RoomMemberEventNames,
+  MemberUpdatedEventNames,
   RTCTrackEventName,
   UserOptions,
   SessionStatus,
